+++ .storybook/main.js
@@ -1,3 +1,4 @@
+import { dirname, join } from 'path';
 module.exports = {
   stories: [
     // Intentionally ordering welcome page first.
@@ -11,8 +12,10 @@ module.exports = {
     // Tooling usage guides.
     '../packages/outline-docs/src/guides/tooling/**/*.mdx',
     // Documentation stories.
-    '../packages/outline-docs/src/guides/**/*.stories.@(js|ts|mdx)',
-    // Component stories.
+    '../packages/outline-docs/src/guides/**/*.@(js|ts|mdx)',
+    // Modern component stories.
+    '../packages/**/docs/*.mdx',
+    // Old component stories.
     '../packages/outline-storybook/stories/**/*.stories.@(js|ts|mdx)',
   ],
   typescript: {
@@ -23,7 +26,7 @@ module.exports = {
     {
       name: '@storybook/addon-essentials',
     },
-    '@storybook/addon-mdx-gfm',
+    getAbsolutePath('@storybook/addon-mdx-gfm'),
   ],
   features: {
     storyStoreV7: true,
@@ -32,7 +35,7 @@ module.exports = {
     modernInlineRender: true,
   },
   framework: {
-    name: '@storybook/web-components-vite',
+    name: getAbsolutePath('@storybook/web-components-vite'),
     options: {},
   },
   docs: {
@@ -40,3 +43,7 @@ module.exports = {
     defaultName: 'Documentation', // set to change the name of generated docs entries
   },
 };
+
+function getAbsolutePath(value) {
+  return dirname(require.resolve(join(value, 'package.json')));
+}