+++ packages/outline-core-link/index.ts
@@ -1,0 +1,3 @@
+export { OutlineCoreLink } from './src/outline-core-link';
+export { default as linkVars } from './src/css-variables/vars-link.css.lit';
+export type { LinkTargetType, LinkRelType } from './src/config';