+++ packages/outline-core-list/index.ts
@@ -1,0 +1,21 @@
+/**
+ * @file Exports for the OutlineCoreList component.
+ *
+ * Any export here is then available via the following import,
+ * which should be available in the consuming application:
+ *
+ * @code
+ * import { OutlineCoreList } from '@phase2/outline-core-list';
+ * import { constantOne, constantTwo } from '@phase2/outline-core-list';
+ * import type { ThingType } from '@phase2/outline-core-list';
+ * @endcode
+ *
+ * Any of the above exports can be used in the consuming application
+ * based on the contents of this file, and creativity of what functionality
+ * you wish to expose to the consuming application(s).
+ */
+
+// Export the primary component.
+export { OutlineCoreList } from './src/outline-core-list';
+
+// Export any additional sub-components, variables and methods as required.