+++ packages/outline-cli/CHANGELOG.md
@@ -1,5 +1,105 @@
 # @phase2/outline-cli
 
+## 0.0.53
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-storybook@0.0.49
+
+## 0.0.52
+
+### Patch Changes
+
+- Hopeful fixes for all the things.
+- Updated dependencies
+  - @phase2/outline-storybook@0.0.48
+  - @phase2/outline-templates@0.0.47
+
+## 0.0.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-storybook@0.0.47
+
+## 0.0.50
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-storybook@0.0.46
+  - @phase2/outline-templates@0.0.46
+
+## 0.0.49
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-storybook@0.0.45
+
+## 0.0.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-storybook@0.0.44
+
+## 0.0.47
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-storybook@0.0.43
+
+## 0.0.46
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-storybook@0.0.42
+  - @phase2/outline-templates@0.0.45
+
+## 0.0.45
+
+### Patch Changes
+
+- @phase2/outline-storybook@0.0.41
+
+## 0.0.44
+
+### Patch Changes
+
+- @phase2/outline-storybook@0.0.40
+
+## 0.0.43
+
+### Patch Changes
+
+- @phase2/outline-storybook@0.0.39
+
+## 0.0.42
+
+### Patch Changes
+
+- Updated dependencies [c33b7e4]
+  - @phase2/outline-templates@0.0.44
+
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-templates@0.0.43
+
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [2cd42bd]
+  - @phase2/outline-storybook@0.0.37
+  - @phase2/outline-templates@0.0.42
+
 ## 0.0.39
 
 ### Patch Changes