+++ packages/outline-docs/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @phase2/outline-docs
 
+## 0.0.18
+
+### Patch Changes
+
+- Updates for light dom package
+- Updated dependencies
+  - @phase2/outline-core@0.2.4
+
+## 0.0.17
+
+### Patch Changes
+
+- Hopeful fixes for all the things.
+- Updated dependencies
+  - @phase2/outline-core@0.2.3
+
+## 0.0.16
+
+### Patch Changes
+
+- Fixes
+
+## 0.0.15
+
+### Patch Changes
+
+- Updates for build
+
+## 0.0.14
+
+### Patch Changes
+
+- Attempted update to docs.
+
+## 0.0.13
+
+### Patch Changes
+
+- c889a22: Update exports
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-core@0.2.0
+
 ## 0.0.11
 
 ### Patch Changes