+++ packages/outline-form/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @phase2/outline-form
 
+## 0.1.2
+
+### Patch Changes
+
+- Updates for light dom package
+- Updated dependencies
+  - @phase2/outline-core@0.2.4
+
+## 0.1.1
+
+### Patch Changes
+
+- Hopeful fixes for all the things.
+- Updated dependencies
+  - @phase2/outline-core@0.2.3
+
+## 0.1.0
+
+### Minor Changes
+
+- Updated lightDomStyles functionality for components
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-core@0.2.0
+
 ## 0.0.8
 
 ### Patch Changes