+++ packages/outline-link/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @phase2/outline-link
 
+## 0.1.6
+
+### Patch Changes
+
+- Hopeful fixes for all the things.
+- Updated dependencies
+  - @phase2/outline-core@0.2.3
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-core@0.2.0
+
 ## 0.1.4
 
 ### Patch Changes