+++ packages/outline-storybook/CHANGELOG.md
@@ -1,5 +1,106 @@
 # @phase2/outline-storybook
 
+## 0.0.49
+
+### Patch Changes
+
+- Updates for light dom package
+- Updated dependencies
+  - @phase2/outline-core@0.2.4
+  - @phase2/outline-docs@0.0.18
+
+## 0.0.48
+
+### Patch Changes
+
+- Hopeful fixes for all the things.
+- Updated dependencies
+  - @phase2/outline-alert@0.1.8
+  - @phase2/outline-code-block@0.1.4
+  - @phase2/outline-docs@0.0.17
+  - @phase2/outline-core@0.2.3
+  - @phase2/outline-static-assets@0.0.4
+
+## 0.0.47
+
+### Patch Changes
+
+- Fixes
+- Updated dependencies
+  - @phase2/outline-core@0.2.2
+
+## 0.0.46
+
+### Patch Changes
+
+- Fixes to CSS tools.
+- Updated dependencies
+  - @phase2/outline-core@0.2.1
+
+## 0.0.45
+
+### Patch Changes
+
+- Fixes
+- Updated dependencies
+  - @phase2/outline-docs@0.0.16
+
+## 0.0.44
+
+### Patch Changes
+
+- Fixes
+
+## 0.0.43
+
+### Patch Changes
+
+- Updates for build
+- Updated dependencies
+  - @phase2/outline-docs@0.0.15
+
+## 0.0.42
+
+### Patch Changes
+
+- Updates to setup and defaults
+
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-docs@0.0.14
+
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [c889a22]
+  - @phase2/outline-docs@0.0.13
+
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-core@0.2.0
+  - @phase2/outline-alert@0.1.7
+  - @phase2/outline-code-block@0.1.3
+  - @phase2/outline-docs@0.0.12
+
+## 0.0.38
+
+### Patch Changes
+
+- Update `outline-core-link` to accomodate slotted links.
+
+## 0.0.37
+
+### Patch Changes
+
+- 2cd42bd: Initial setup of outline-core-link
+
 ## 0.0.36
 
 ### Patch Changes