+++ packages/outline-style-guide/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @phase2/outline-style-guide
 
+## 0.0.6
+
+### Patch Changes
+
+- Hopeful fixes for all the things.
+- Updated dependencies
+  - @phase2/outline-core@0.2.3
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @phase2/outline-core@0.2.0
+
 ## 0.0.4
 
 ### Patch Changes